Port Chunk System Tile Map demo from the legacy sandbox to Storybook
See also: #1348
