--- sandbox/html/chunk-system-tile-map-demo.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Chunk system tile map demo</title>
-
-    <style type="text/css">
-      html {
-        background-color: #345;
-      }
-
-      #game {
-        margin: auto;
-        display: block;
-      }
-    </style>
-
-    <link rel="icon" type="image/x-icon" href="../images/favicon.png" />
-  </head>
-  <body>
-    <div id="container" class="container">
-      <canvas id="game"></canvas>
-    </div>
-    <script type="text/javascript" src="../lib/excalibur.js"></script>
-    <script type="text/javascript" src="../src/chunk-system-tile-map-demo.js"></script>
-  </body>
-</html>
