+++ src/stories/assets/tiles.png
